Fix normal functions not listed in profile, mostly on Linux
Linux is weird. The simple check against OP_PROC didn't work there so it now tests if the two frames are the same when entering a "normal" function. This is to prevent public functions from being referenced twice in the list of function (one time as normal, and one as public).

There also was a typo(?) in Profiler::Debug:

if (call_stack_.GetTop().function().type() == Function::PUBLIC)

should be

if (call_stack_.GetTop().function().type() != Function::PUBLIC)
                                           ^
which is equivalent to

if (call_stack_.GetTop().function().type() == Function::NORMAL)

so I used the last one (looks clearer)