Remove redundant fractional parts in JSON output
Due to the use of std::fixed all integer numbers had trailing
".0000". Also removed std::setprecision manipulators - the output
of the JSON writer is not supposed to be read by humans directly
(because there are HTML and text writers for that purpose) but
rather fed into some sort of parser that may perform various
computations on the numbers present there, and truncating them
can have negative effects on the results.