Attach profiler on demand
Always register AMX natives regardless of whether the script is going
to be profiled or not. This allows for dynamic attachment of the profiler
on the first call to Profiler_Start().

Replace the profile_gamemode variable with profiler_gamemodes which takes
a list of gamemodes to be profiled. This makes it slightly easier to
switch between different gamemodes without worrying about disabling
or enabling profiler in server.cfg each time (more fine-grained control).

And finally, replace "profile" in other config variables with "profiler"
to indicate that they belong to the profiler plugin. Some variables didn't
have a prefix at all (i.e. call_graph) which was a little confusing.