Improve file path comparison algorithm
Make the file path equivalence comparison more accurate. Instead of simply comparing file
paths check if the two paths refer to the same node on the file system.

The previous implementation didn't work in some cases, for example, when AMX_PATH was set
to an absolute path such as C:\samp-server\gamemodes\somedir.